Refactor to share code between reference and optimized kernels.
The goal is to reduce code duplication, improve readability of the
optimized kernels and reduce the maintenance overhead.
