+++ tensorflow/lite/micro/kernels/reduce_test.cc
@@ -13,6 +13,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
+#include "tensorflow/lite/micro/kernels/reduce.h"
+
 #include "tensorflow/lite/c/builtin_op_data.h"
 #include "tensorflow/lite/c/common.h"
 #include "tensorflow/lite/micro/all_ops_resolver.h"
@@ -99,7 +101,7 @@ void TestMeanFloatInput4D(int* input_dims_data, const float* input_data,
   TfLiteIntArray* output_dims = IntArrayFromInts(output_dims_data);
   const int output_dims_count = ElementCount(*output_dims);
 
-  const TfLiteRegistration registration = tflite::ops::micro::Register_MEAN();
+  const TfLiteRegistration registration = tflite::Register_MEAN();
 
   constexpr int num_of_inputs = 2;   // input and axis
   constexpr int num_of_outputs = 1;  // output
@@ -218,7 +220,7 @@ void TestMeanOpQuantized(int* input_dims_data, const float* input_data,
       kTfLiteOk,
       ValidateReduceGoldens(tensors, tensors_size, expected_output_data_quant,
                             output_data_quant, output_dims_count,
-                            tflite::ops::micro::Register_MEAN(), params, 1.0));
+                            tflite::Register_MEAN(), params, 1.0));
 }
 
 }  // namespace
@@ -515,7 +517,7 @@ TF_LITE_MICRO_TEST(FloatMaxOpTestNotKeepDims) {
 
   tflite::testing::TestReduceOpFloat(
       input_shape, input_data, axis_shape, axis_data, output_shape, output_data,
-      expected_output_data, tflite::ops::micro::Register_REDUCE_MAX(), &params);
+      expected_output_data, tflite::Register_REDUCE_MAX(), &params);
 }
 
 TF_LITE_MICRO_TEST(FloatMaxOpTestKeepDims) {
@@ -533,7 +535,7 @@ TF_LITE_MICRO_TEST(FloatMaxOpTestKeepDims) {
 
   tflite::testing::TestReduceOpFloat(
       input_shape, input_data, axis_shape, axis_data, output_shape, output_data,
-      expected_output_data, tflite::ops::micro::Register_REDUCE_MAX(), &params);
+      expected_output_data, tflite::Register_REDUCE_MAX(), &params);
 }
 
 TF_LITE_MICRO_TEST(Int8MaxOpTestKeepDims) {
@@ -557,7 +559,7 @@ TF_LITE_MICRO_TEST(Int8MaxOpTestKeepDims) {
       input_shape, input_data, input_data_quant, input_scale, input_zp,
       axis_shape, axis_data, output_shape, expected_output_data,
       output_data_quant, expected_output_data_quant, input_scale, input_zp,
-      tflite::ops::micro::Register_REDUCE_MAX(), &params);
+      tflite::Register_REDUCE_MAX(), &params);
 }
 
 TF_LITE_MICRO_TEST(Int8MaxOpTestWithoutKeepDims) {
@@ -583,7 +585,7 @@ TF_LITE_MICRO_TEST(Int8MaxOpTestWithoutKeepDims) {
       input_shape, input_data, input_data_quant, input_scale, input_zp,
       axis_shape, axis_data, output_shape, expected_output_data,
       output_data_quant, expected_output_data_quant, output_scale, output_zp,
-      tflite::ops::micro::Register_REDUCE_MAX(), &params);
+      tflite::Register_REDUCE_MAX(), &params);
 }
 
 TF_LITE_MICRO_TEST(MeanInt84DWithoutKeepDimsWithPrecision) {
